backend/requirements.txt
flask==3.0.0
Flask-CORS==5.0.0
requests==2.31.0
python-dotenv==1.0.0
waitress==2.1.2
minio==7.2.9